requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via django
black==21.6b0
    # via -r requirements.in
click==8.0.1
    # via black
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.5
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
factory-boy==3.2.0
    # via -r requirements.in
faker==8.10.1
    # via factory-boy
importlib-metadata==4.6.1
    # via click
mypy-extensions==0.4.3
    # via black
names==0.3.0
    # via -r requirements.in
pathspec==0.8.1
    # via black
psycopg2-binary==2.9.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
python-dotenv==0.18.0
    # via -r requirements.in
pytz==2021.1
    # via django
regex==2021.7.6
    # via black
six==1.16.0
    # via python-dateutil
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via black
typed-ast==1.4.3
    # via black
typing-extensions==3.10.0.0
    # via
    #   asgiref
    #   black
    #   importlib-metadata
zipp==3.5.0
    # via importlib-metadata
